[Enhancement][Optimized] Optimized the OrderCode to Generic
