+++ LapStopApiSolution/Logic/Repositories/Extensions/RepositoryEmployeeExtension.cs
@@ -1,5 +1,6 @@
 ﻿using Domains.Models;
 using DTO.Input.FromQuery.Parameters;
+using Repositories.Extensions.Utility;
 using System.Linq.Dynamic.Core;
 using System.Reflection;
 using System.Text;
@@ -33,40 +34,9 @@ namespace Repositories.Extensions
             {
                 return employees.OrderBy(e => e.EmployeeCode);  // DEFAULT
             }
+
+            string orderQuery = OrderQueryBuilder.CreateOrderQueryFor<Employee>(parameter.OrderBy);
             
-            // Get all properties of Employee
-            var propertyInfos = typeof(Employee).GetProperties(BindingFlags.Public | BindingFlags.Instance);
-            
-            // "firstName, lastName desc" -> [firstName], [lastName]
-            var orderParams = parameter.OrderBy.Split(',');
-
-            var orderQueryBuilder = new StringBuilder();
-            foreach (var param in orderParams)
-            {
-                if (string.IsNullOrWhiteSpace(param))   // " , lastName desc" -> [''], [lastName]
-                {
-                    continue;
-                }
-
-                // "firstName" --> "firstName"
-                // "lastName desc" --> "lastName"
-                var propertyNameFromQuery = param.Split(" ")[0];
-
-                var objectProperty = propertyInfos.FirstOrDefault(
-                                        pi => pi.Name.Equals(propertyNameFromQuery, 
-                                                             StringComparison.InvariantCultureIgnoreCase));
-                if (objectProperty == null) // Invalid PropName
-                {
-                    continue;
-                }
-
-                // default -> ascending
-                var direction = param.EndsWith(" desc") ? "descending" : "ascending";
-                orderQueryBuilder.Append($"{objectProperty.Name.ToString()} {direction},");
-                // "firstName, lastName desc" --> "firstName ascending, lastName descending,"
-            }
-
-            var orderQuery = orderQueryBuilder.ToString().TrimEnd(',', ' ');
             if (string.IsNullOrWhiteSpace(orderQuery))          // check NULL
             {
                 return employees.OrderBy(e => e.EmployeeCode);  // DEFAULT - OrderBy of LINQ
